chore: move timestats back to metrics
