+++ runner/internal/metrics/aggregate_test.go
@@ -5,8 +5,8 @@ import (
 	"testing"
 	"time"
 
-	"github.com/benchttp/engine/internal/timestats"
 	"github.com/benchttp/engine/runner/internal/metrics"
+	"github.com/benchttp/engine/runner/internal/metrics/timestats"
 	"github.com/benchttp/engine/runner/internal/recorder"
 )
