+++ runner/internal/metrics/timestats/sort.go
